BUG: Ensure that we view the datetime64s in the index as 64bit
Using int with 32bit numpy produces an array of 2n ints, each
representing half a datetime.

+++ zipline/assets/asset_writer.py
@@ -475,7 +475,7 @@ class AssetDBWriter(with_metaclass(ABCMeta)):
         except TypeError:
             index = index.tz_convert('UTC')
 
-        return index.view(int)
+        return index.view(np.int64)
 
     @abstractmethod
     def _load_data(self):
